Improved Projectile Interface, added static_cast, changed magic variables to constexpr

+++ Source/Player.cpp
@@ -12,7 +12,7 @@ void Player::UpdateMovement()
 	{
 		movement = 1;
 	}
-	x_pos = std::clamp(x_pos += SPEED * movement, PLAYER_RADIUS, GetScreenWidth() - PLAYER_RADIUS);
+	x_pos = std::clamp(x_pos += SPEED * movement, PLAYER_RADIUS, static_cast<float>(GetScreenWidth()) - PLAYER_RADIUS);
 }
 void Player::UpdateAnimation() noexcept
 {
@@ -32,6 +32,6 @@ void Player::Update()
 void Player::Render(Texture2D texture) noexcept
 {
 	const Rectangle src = { 0,0,static_cast<float>(texture.width), static_cast<float>(texture.height) };
-	const Rectangle dst = { x_pos, GetScreenHeight() - PLAYER_BASE_HEIGHT, 100, 100 };
+	const Rectangle dst = { x_pos, static_cast<float>(GetScreenHeight()) - PLAYER_BASE_HEIGHT, 100, 100 };
 	DrawTexturePro(texture, src, dst, PosOr, 0, WHITE);
 }
